add abstract circuit
