+++ examples/hchainhamiltonian.py
@@ -4,6 +4,7 @@ Get molecule qubit format Hamiltonian from openfermion.
 
 import time
 import tensorflow as tf
+from scipy import sparse
 from openfermion.chem import MolecularData
 from openfermion.transforms import (
     get_fermion_operator,
@@ -14,7 +15,6 @@ from openfermion.transforms import (
 )
 from openfermion.utils import up_then_down
 from openfermionpyscf import run_pyscf
-from scipy import sparse
 
 import tensorcircuit as tc
