+++ examples/vqetfim_benchmark.py
@@ -115,7 +115,7 @@ hamiltonian_sparse = K.coo_sparse_matrix(
         np.stack([hamiltonian_sparse_numpy.row, hamiltonian_sparse_numpy.col])
     ),
     hamiltonian_sparse_numpy.data,
-    shape=(2**n, 2**n),
+    shape=(2 ** n, 2 ** n),
 )
 
 if enable_dense is True:
