+++ tensorcircuit/applications/utils.py
@@ -54,11 +54,11 @@ def amplitude_encoding(
     norm = tf.linalg.norm(fig, axis=1)
     norm = norm[..., tf.newaxis]
     fig = fig / norm
-    if fig.shape[1] < 2**qubits:
+    if fig.shape[1] < 2 ** qubits:
         fig = tf.concat(
             [
                 fig,
-                tf.zeros([fig.shape[0], 2**qubits - fig.shape[1]], dtype=tf.float64),
+                tf.zeros([fig.shape[0], 2 ** qubits - fig.shape[1]], dtype=tf.float64),
             ],
             axis=1,
         )
@@ -205,7 +205,7 @@ def train_qml_vag(
     with tf.GradientTape() as tape:
         tape.watch(nnp)
         cnnp = tf.cast(nnp, dtype=tf.complex64)
-        c = Circuit(nqubits, inputs=np.ones([1024], dtype=np.complex64) / 2**5)
+        c = Circuit(nqubits, inputs=np.ones([1024], dtype=np.complex64) / 2 ** 5)
         for epoch in range(epochs):
             for i in range(nqubits):
                 c.rz(i, theta=cnnp[3 * epoch, i])  # type: ignore
@@ -275,7 +275,7 @@ def validate_qml_vag(
 ) -> Any:
     xs, ys = gdata
     cnnp = tf.cast(nnp, dtype=tf.complex64)
-    c = Circuit(nqubits, inputs=np.ones([1024], dtype=np.complex64) / 2**5)
+    c = Circuit(nqubits, inputs=np.ones([1024], dtype=np.complex64) / 2 ** 5)
     for epoch in range(epochs):
         for i in range(nqubits):
             c.rz(i, theta=cnnp[3 * epoch, i])  # type: ignore
@@ -401,7 +401,7 @@ def TFIM1Denergy(
         Jzz *= 4
     for i in range(L):
         q = np.pi * (2 * i - (1 + (-1) ** L) / 2) / L
-        e -= np.abs(Jx) / 2 * np.sqrt(1 + Jzz**2 / 4 / Jx**2 - Jzz / Jx * np.cos(q))
+        e -= np.abs(Jx) / 2 * np.sqrt(1 + Jzz ** 2 / 4 / Jx ** 2 - Jzz / Jx * np.cos(q))
     return e
 
 
@@ -413,7 +413,7 @@ def Heisenberg1Denergy(L: int, Pauli: bool = True, maxiters: int = 1000) -> floa
     phi2 = np.zeros([L // 2, L // 2])
     lamb = np.array([2 * i + 1 for i in range(L // 2)])
     for _ in range(maxiters):
-        k = 1 / L * (2 * np.pi * lamb + np.sum(phi, axis=-1) - np.diag(phi))
+        k = 1 / L * (2 * np.pi * lamb + np.sum(phi, axis=-1) - np.diag(phi))  # type: ignore
         for i in range(L // 2):
             for j in range(L // 2):
                 phi2[i, j] = (
@@ -427,7 +427,7 @@ def Heisenberg1Denergy(L: int, Pauli: bool = True, maxiters: int = 1000) -> floa
                     )
                     * 2
                 )
-        if np.allclose(phi, phi2, rtol=error):  # converged
+        if np.allclose(phi, phi2, rtol=error):  # type: ignore # converged
             break
         phi = phi2.copy()
     else:
