+++ tensorcircuit/translation.py
@@ -40,8 +40,8 @@ def perm_matrix(n: int) -> Tensor:
     :return: The permutation matrix P
     :rtype: Tensor
     """
-    p_mat = np.zeros([2**n, 2**n])
-    for i in range(2**n):
+    p_mat = np.zeros([2 ** n, 2 ** n])
+    for i in range(2 ** n):
         bit = i
         revs_i = 0
         for j in range(n):
