+++ tests/test_quantum.py
@@ -319,7 +319,7 @@ def test_heisenberg_ham(tfb):
 @pytest.mark.parametrize("backend", [lf("npb"), lf("tfb"), lf("jaxb")])
 def test_reduced_density_from_density(backend):
     n = 6
-    w = np.random.normal(size=[2**n]) + 1.0j * np.random.normal(size=[2**n])
+    w = np.random.normal(size=[2 ** n]) + 1.0j * np.random.normal(size=[2 ** n])
     w /= np.linalg.norm(w)
     rho = np.reshape(w, [-1, 1]) @ np.reshape(np.conj(w), [1, -1])
     dm1 = tc.quantum.reduced_density_matrix(w, cut=[0, 2])
@@ -328,9 +328,9 @@ def test_reduced_density_from_density(backend):
 
     # with p
     n = 5
-    w = np.random.normal(size=[2**n]) + 1.0j * np.random.normal(size=[2**n])
+    w = np.random.normal(size=[2 ** n]) + 1.0j * np.random.normal(size=[2 ** n])
     w /= np.linalg.norm(w)
-    p = np.random.normal(size=[2**3])
+    p = np.random.normal(size=[2 ** 3])
     p = tc.backend.softmax(p)
     p = tc.backend.cast(p, "complex128")
     rho = np.reshape(w, [-1, 1]) @ np.reshape(np.conj(w), [1, -1])
@@ -342,7 +342,7 @@ def test_reduced_density_from_density(backend):
 @pytest.mark.parametrize("backend", [lf("npb"), lf("tfb"), lf("jaxb")])
 def test_mutual_information(backend):
     n = 5
-    w = np.random.normal(size=[2**n]) + 1.0j * np.random.normal(size=[2**n])
+    w = np.random.normal(size=[2 ** n]) + 1.0j * np.random.normal(size=[2 ** n])
     w /= np.linalg.norm(w)
     rho = np.reshape(w, [-1, 1]) @ np.reshape(np.conj(w), [1, -1])
     dm1 = tc.quantum.mutual_information(w, cut=[1, 2, 3])
